#3 rename: 컴포넌트 디렉토리 수정
